+++ src/renderer/quad/draw_quad.cpp
@@ -5,7 +5,7 @@
 #include <glm/glm.hpp>
 #include <glm/gtx/rotate_vector.hpp>
 
-void QuadRenderer::draw(const Quad &quad, const int layer) {
+void QuadRenderer::draw(const Quad &quad, const float layer) {
     if (m_buffer_index + 6 > m_max_buffer_size) {
         flush();
     }
